Refactor CommunityTransactions component by consolidating transaction-related hooks and removing unused files for improved clarity and maintainability.
