+++ packages/commonwealth/client/scripts/state/api/communityStake/getFeeManagerBalance.ts
@@ -1,8 +1,8 @@
 import { useQuery } from '@tanstack/react-query';
+import { useNamespaceFactory } from 'views/pages/CreateCommunity/steps/CommunityTransactions';
 
 import NamespaceFactory from 'helpers/ContractHelpers/NamespaceFactory';
 import { ContractMethods } from 'state/api/config';
-import { useNamespaceFactory } from 'views/pages/CreateCommunity/steps/CommunityTransactions';
 
 const GET_FEE_MANAGER_BALANCE_STALE_TIME = 60 * 1_000; // 60 sec
