+++ packages/commonwealth/client/scripts/views/pages/CreateCommunity/steps/CommunityTransactions/SignCommunityTransactions/SignCommunityTransactions.scss
@@ -1,4 +1,4 @@
-@import '../../../../../styles/shared';
+@import '../../../../../../styles/shared';
 
 .SignCommunityTransactions {
   display: flex;
