+++ packages/commonwealth/client/scripts/views/pages/CreateCommunity/steps/CommunityTransactions/SignCommunityTransactions/index.ts
@@ -1,0 +1,3 @@
+import SignCommunityTransactions from './SignCommunityTransactions';
+
+export default SignCommunityTransactions;
