+++ packages/commonwealth/client/scripts/views/pages/CreateCommunity/steps/CommunityTransactions/helpers/transactionUtils.ts
@@ -1,4 +1,4 @@
-import { TransactionConfig, TransactionHookResult } from './types';
+import { TransactionConfig, TransactionHookResult } from '../types';
 
 /**
  * Creates a namespace transaction configuration
