+++ packages/commonwealth/client/scripts/views/pages/CreateCommunity/steps/CommunityTransactions/helpers/useNamespaceFactory.ts
