+++ packages/commonwealth/client/scripts/views/pages/CreateCommunity/steps/CommunityTransactions/helpers/useStakeTransaction.tsx
@@ -12,7 +12,7 @@ import {
   TransactionData,
   TransactionHookResult,
   defaultTransactionState,
-} from './types';
+} from '../types';
 import useNamespaceFactory from './useNamespaceFactory';
 
 interface UseStakeTransactionProps {
