+++ packages/commonwealth/client/scripts/views/pages/CreateCommunity/steps/CommunityTransactions/types.ts
@@ -1,4 +1,3 @@
-import AddressInfo from 'models/AddressInfo';
 import { ActionStepProps } from 'views/pages/CreateCommunity/components/ActionSteps/types';
 
 // Transaction state types
@@ -34,18 +33,6 @@ export interface TransactionHookResult {
   action: () => Promise<void>;
 }
 
-// SignCommunityTransactions props
-export interface SignCommunityTransactionsProps {
-  title: string;
-  description: string;
-  transactions: TransactionConfig[];
-  isPreventLeaveEnabled?: boolean;
-  backButton?: {
-    label: string;
-    action: () => void;
-  };
-}
-
 // Community data structure
 export type CommunityData = {
   symbol: string;
@@ -61,35 +48,6 @@ export type StakeData = {
   symbol: string;
   namespace: string;
 };
-
-export interface SignStakeTransactionsProps {
-  communityStakeData: StakeData;
-  selectedAddress: AddressInfo;
-  createdCommunityId: string;
-  chainId: string;
-  onlyNamespace?: boolean;
-  hasNamespaceReserved?: boolean;
-  onReserveNamespaceSuccess: () => void;
-  onLaunchStakeSuccess: () => void;
-  backButton?: {
-    label: string;
-    action: () => void;
-  };
-}
-
-export interface EnableStakeProps {
-  communityStakeData: StakeData;
-  chainId: string;
-  onlyNamespace?: boolean;
-  backButton?: {
-    label: string;
-    action: () => void;
-  };
-  confirmButton?: {
-    label: string;
-    action: (data: StakeData) => void;
-  };
-}
 
 export const defaultActionState: ActionState = {
   state: 'not-started',
